fix(uploads,theme): cast GD dimensions to int; guard localStorage quota
ImageUploader::processImage was passing float $targetWidth/$targetHeight
to imagecreatetruecolor() and imagecopyresampled(), triggering PHP 8.2
deprecation warnings on the admin partner-logo upload path. Cast to int
at the call sites and at the computed offsets so float arithmetic stays
local to the ratio logic.

ThemeContext::storeTheme / storePreferences could throw QuotaExceededError
on browsers with full localStorage, propagating to Sentry as an unhandled
rejection from the /blog page. Wrap setItem in try/catch — the in-memory
theme state is the source of truth, persistence is best-effort.

Fixes NEXUS-PHP-19
Fixes NEXUS-PHP-1A
Fixes NEXUS-REACT-7

+++ app/Core/ImageUploader.php
@@ -268,32 +268,32 @@ class ImageUploader
                 $newHeight = (int)($srcWidth / $thumbRatio);
             }
 
-            $xOffset = ($srcWidth - $newWidth) / 2;
-            $yOffset = ($srcHeight - $newHeight) / 2;
+            $xOffset = (int) (($srcWidth - $newWidth) / 2);
+            $yOffset = (int) (($srcHeight - $newHeight) / 2);
 
-            $thumb = \imagecreatetruecolor($targetWidth, $targetHeight);
+            $thumb = \imagecreatetruecolor((int) $targetWidth, (int) $targetHeight);
 
             if ($mime == 'image/png' || $mime == 'image/webp') {
                 \imagealphablending($thumb, false);
                 \imagesavealpha($thumb, true);
             }
 
-            \imagecopyresampled($thumb, $image, 0, 0, $xOffset, $yOffset, $targetWidth, $targetHeight, $newWidth, $newHeight);
+            \imagecopyresampled($thumb, $image, 0, 0, $xOffset, $yOffset, (int) $targetWidth, (int) $targetHeight, $newWidth, $newHeight);
             $image = $thumb;
         } elseif (isset($options['width'])) {
             $ratio = $srcWidth / $srcHeight;
             if ($targetWidth / $targetHeight > $ratio) {
-                $targetWidth = $targetHeight * $ratio;
+                $targetWidth = (int) ($targetHeight * $ratio);
             } else {
-                $targetHeight = $targetWidth / $ratio;
+                $targetHeight = (int) ($targetWidth / $ratio);
             }
 
-            $thumb = \imagecreatetruecolor($targetWidth, $targetHeight);
+            $thumb = \imagecreatetruecolor((int) $targetWidth, (int) $targetHeight);
             if ($mime == 'image/png' || $mime == 'image/webp') {
                 \imagealphablending($thumb, false);
                 \imagesavealpha($thumb, true);
             }
-            \imagecopyresampled($thumb, $image, 0, 0, 0, 0, $targetWidth, $targetHeight, $srcWidth, $srcHeight);
+            \imagecopyresampled($thumb, $image, 0, 0, 0, 0, (int) $targetWidth, (int) $targetHeight, $srcWidth, $srcHeight);
             $image = $thumb;
         }
